Switch between active and passive mixing
This change adds a toggle in the menu bar to
switch back and forth between active and
passive DJ styles.

When actively mixing between multiple decks, it is
preferable to use the 'oldest' deck. But when using
autoplay (e.g., more passively mixing) or only a single
deck, it is preferable to always use the 'newest'
deck.

* added a new mixmode global :(
* updated the comments to reflect algo changes
* updated the readme to document the menu
* put some separators in between the menu items to
  hopefuly prevent accidents...